feat(ws-handler): adds command and instructions on how to use it
This command handles when the scratchpad workspace gets the focus (which shouldn't happen). It will move focus back to the last focused workspace and take the focused window to that workspace too.
This allow you to use different tools to focus windows in scratchpad, like notifications, external launchers, etc., and behave as "summoning" the window to the current workspace instead of focusing the scratchpad workspace.

Add this snippet in your `~/.aerospace.toml` config:
```toml
exec-on-workspace-change = ['/bin/bash', '-c',
    'aerospace-scratchpad workspace-handler $AEROSPACE_FOCUSED_WORKSPACE'
]
